Just remove settings for now since we don't have a gui
